requirements.txt
flask
requests
sqlalchemy
psycopg[binary]
python-dotenv
